--- src/test/java/seedu/address/testutil/TypicalIndexes.java
@@ -1,12 +1,0 @@
-//package seedu.address.testutil;
-//
-//import seedu.address.commons.core.index.Index;
-//
-///**
-// * A utility class containing a list of {@code Index} objects to be used in tests.
-// */
-//public class TypicalIndexes {
-//    public static final Identification INDEX_FIRST_PERSON = Identification.fromOneBased(1);
-//    public static final Identification INDEX_SECOND_PERSON = Identification.fromOneBased(2);
-//    public static final Identification INDEX_THIRD_PERSON = Identification.fromOneBased(3);
-//}
